requirements.txt
# Core Python packages
anyio==4.6.2.post1
APScheduler==3.10.4
python-dotenv==1.0.1

# Database packages
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

# Telegram Bot API
python-telegram-bot==21.9

# Timezone handling
pytz==2024.2

# Firebase (if using Firestore)
firebase-admin==6.5.0
google-cloud-firestore==2.19.0

# Additional dependencies
idna==3.10
sniffio==1.3.1
typing-extensions==4.12.2

# Development and testing
unittest-xml-reporting==3.2.0
